+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2023-090-07
+## [Unreleased] - 2023-09-08
+
+### Added
+ - Add `WComponent` class as Livewire component with support of `WAccount`. 
+
+## [Unreleased] - 2023-09-07
 
 ### Added
  - Add `url` in `initializeIndex` as parameter for custom usages.